+++ semcore/color-picker/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [2.34.2] - 2024-06-07
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/button` [5.28.1 ~> 5.29.0], `@semcore/input` [4.28.1 ~> 4.28.2]).
+
 ## [2.34.1] - 2024-05-28
 
 ### Changed